Cast goodRxns to double; precompute with nSamples=0 in seed test
ismember returns its index output as a non-double type in some MATLAB
versions; randperm is stricter than randi and rejects non-double scalar
inputs with MATLAB:randperm:inputType. Add explicit double() cast to both
conversion sites (incoming and outgoing goodRxns).

Change the goodRxns precomputation call in randomSamplingDeterministicWithSeed
from nSamples=5 to nSamples=0. With nSamples=0 the function runs the full
goodRxns detection (LP-based) but skips the sampling loop, so the
precomputation costs one detection pass instead of one detection + five
sample LPs. The two seeded calls then skip detection entirely and only
run five sample LPs each.

+++ analysis/randomSampling.m
@@ -110,7 +110,7 @@ model=simplifyModel(model,false,false,true,true);
 %loops below, dropping any reactions that were removed by simplification.
 if ~isempty(goodRxns)
     [tf, simpIdx] = ismember(originalRxns(goodRxns), model.rxns);
-    goodRxns = simpIdx(tf);
+    goodRxns = double(simpIdx(tf));
 end
 
 %Check that the model is feasible given the constraints
@@ -222,5 +222,5 @@ else
 end
 %Return goodRxns as original-model indices so callers can pass them back
 %safely across separate calls (which each re-simplify the model internally).
-goodRxns = simpToOrig(goodRxns);
+goodRxns = double(simpToOrig(goodRxns));
 end
